Namespace definition grammar fix (#254)
* Fix the `namespace-definition` production

The `namespace-definition` production was changed to use `name` instead
of `namespace-name` in c1ac5314798f6740f and a6520fb35b746a6e6c275. This
appears to be a mistake, since `name` can not produce a `\`.

The spec was originally using `namespace-name` for
`namespace-definition`. This commit restores that grammar production.

* Regenerate using `tools/pre-commit`

It appears that the pre-commit scripts were not used to properly
regenerate the spec recently.

This caused some small styling changes (using html entities for some
chars).

It also caused a bigger change due to `print` being moved out of the
intrinsics section in 63d72a6df5d2c7b1b7a8a. It's not completely clear
if this was intentional or not, since `print` is a language intrinsic.